+++ NEWS.md
@@ -6,7 +6,8 @@
 
 ## Changes to functions
 
-* Added `iterations` argument to `get_predicted` to control the number of draws returned for Bayesian models.
+* Added `get_predicted()` for Bayesian models returns a data.frame by default instead of a matrix.
+* Added `iterations` argument to `get_predicted()` to control the number of draws returned for Bayesian models.
 
 # insight 0.13.1